refactor: 불필요한 메서드 삭제, generateStackView(about:) 메서드 분리
