+++ Calculator/Calculator/view/CalculateItemStackView.swift
@@ -16,38 +16,16 @@ final class CalculateItemStackView: UIStackView {
                                           roundingMode: .halfUp,
                                           usesSignificantDigits: true,
                                           maximumSignificantDigits: 20)
+    private let viewGenerator = ViewGenerator()
     
     func removeAllSubviews() {
         self.subviews.forEach { $0.removeFromSuperview() }
     }
     
     func add(_ currentItem: CurrentItem) {
-        guard let subview = generateStackView(about: currentItem) else { return }
+        guard let subview = viewGenerator.generateStackView(about: currentItem) else { return }
         
-        add(subview, to: self)
+        self.addArrangedSubview(subview)
         calculateItemScrollView.didAddSubview(subview)
-    }
-    
-    private func generateStackView(about currentItem: CurrentItem) -> UIStackView? {
-        let operand = UILabel()
-        operand.text = numberFormatter.convertToDecimal(from: currentItem.operand)
-        operand.textColor = UIColor.white
-        operand.font = UIFont.preferredFont(forTextStyle: .title3)
-        
-        let `operator` = UILabel()
-        `operator`.text = currentItem.operator
-        `operator`.textColor = UIColor.white
-        `operator`.font = UIFont.preferredFont(forTextStyle: .title3)
-        
-        let result = UIStackView()
-        result.addArrangedSubview(`operator`)
-        result.addArrangedSubview(operand)
-        result.spacing = 8
-        
-        return result
-    }
-    
-    private func add(_ subview: UIStackView, to superview: UIStackView) {
-        superview.addArrangedSubview(subview)
     }
 }
